api: Use AsRef<Path> instead of &Path
This matches the convention used in std and makes the API simpler
when calling open() as you can just provide a &str or String and
it will be converted into a &Path autmmatically.  This change is
backwards compatible, so a full semver bump should not be required.

You can see this pattern in std in a number of places:
* https://doc.rust-lang.org/std/fs/struct.File.html#method.open
* https://doc.rust-lang.org/std/fs/fn.read_dir.html
* ... (basically everything that uses a path) ...