Read-only and read-write Database and Transaction
This commit unifies the `Transaction` and `ReadonlyTransaction` types
into a single `Transaction` type that is generic with respect to whether
the transaction is read-only or read-write. There are two new type
definitions, `ReadOnlyTransaction` and `ReadWriteTransaction`, which
should provide some syntactic sugar for applications.

Similarly, this commit changes the `Database` type to be generic with
respect to read-only vs read-write access, as well as creating two new
type definitions, `ReadOnlyDatabase` and `ReadWriteDatabase`. Read-only
databases only provide non-mutating methods. For example, it will be a
compiler error for an application to call the `put` method on a
`ReadOnlyDatabase`.